add cross-package Rollup config ES6 module interop file
